commit ver1

+++ HospitalManagement/src/main/java/com/hospital/repository/EmployeeRepositoryCustom.java
@@ -1,14 +1,14 @@
 package com.hospital.repository;
 
+import java.util.List;
+
 import com.hospital.dto.PatientRecordDTO;
 import com.hospital.model.Appointment;
 import com.hospital.model.Examination;
 
-import java.util.List;
-
 public interface EmployeeRepositoryCustom {
-    public List<Examination> findExaminationsByExaminatorId(int employee_id);
-    public List<Examination> findUnfinishedExaminationsByExaminatorId(int employee_id);
-    public List<Appointment> findAppointmentByDoctorId(int employee_id);
+    public List<Examination> findExaminationsByExaminatorId(Long employee_id);
+    public List<Examination> findUnfinishedExaminationsByExaminatorId(Long employee_id);
+    public List<Appointment> findAppointmentByDoctorId(Long employee_id);
     public PatientRecordDTO findPatientHistoryRecord(Long patient_id);
 }
